moving file according to PR notice, deleting duplicate classes
