+++ PhpStormStubsMap.php
@@ -258,17 +258,17 @@ const CLASSES = array (
   'DirectoryIterator' => 'SPL/SPL_c1.php',
   'DivisionByZeroError' => 'Core/Core_c.php',
   'DomainException' => 'SPL/SPL.php',
-  'Ds\\Collection' => 'ds/ds_c.php',
-  'Ds\\Deque' => 'ds/ds_c.php',
-  'Ds\\Hashable' => 'ds/ds_c.php',
-  'Ds\\Map' => 'ds/ds_c.php',
-  'Ds\\Pair' => 'ds/ds_c.php',
-  'Ds\\PriorityQueue' => 'ds/ds_c.php',
-  'Ds\\Queue' => 'ds/ds_c.php',
-  'Ds\\Sequence' => 'ds/ds_c.php',
-  'Ds\\Set' => 'ds/ds_c.php',
-  'Ds\\Stack' => 'ds/ds_c.php',
-  'Ds\\Vector' => 'ds/ds_c.php',
+  'Ds\\Collection' => 'ds/ds.php',
+  'Ds\\Deque' => 'ds/ds.php',
+  'Ds\\Hashable' => 'ds/ds.php',
+  'Ds\\Map' => 'ds/ds.php',
+  'Ds\\Pair' => 'ds/ds.php',
+  'Ds\\PriorityQueue' => 'ds/ds.php',
+  'Ds\\Queue' => 'ds/ds.php',
+  'Ds\\Sequence' => 'ds/ds.php',
+  'Ds\\Set' => 'ds/ds.php',
+  'Ds\\Stack' => 'ds/ds.php',
+  'Ds\\Vector' => 'ds/ds.php',
   'EmptyIterator' => 'SPL/SPL.php',
   'Error' => 'Core/Core_c.php',
   'ErrorException' => 'Core/Core_c.php',
